requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.12.1
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
babel==2.18.0
bangla==0.0.5
blinker==1.9.0
blis==1.3.3
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.13
Cython==3.2.4
dateparser==1.1.8
decorator==5.2.1
demucs==4.0.1
docopt==0.6.2
dora_search==0.1.12
einops==0.8.2
encodec==0.1.1
exceptiongroup==1.3.1
filelock==3.20.0
Flask==3.1.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
g2pkk==0.1.2
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.23.2
HyperPyYAML==1.2.3
idna==3.11
importlib_metadata==8.7.1
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.5.3
jsonlines==1.2.0
julius==0.2.7
kiwisolver==1.4.9
lameenc==1.8.1
lazy_loader==0.4
librosa==0.10.0
lightning==2.6.1
lightning-utilities==0.15.3
llvmlite==0.46.0
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
murmurhash==1.0.15
networkx==2.8.8
nltk==3.9.3
num2words==0.5.14
numba==0.64.0
numpy==1.24.3
omegaconf==2.3.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
openunmix==1.3.0
optuna==4.7.0
packaging==26.0
pandas==1.5.3
pillow==12.0.0
platformdirs==4.9.2
pooch==1.9.0
preshed==3.0.12
primePy==1.3
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyannoteai-sdk==0.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pynndescent==0.6.0
pyparsing==3.3.2
pypinyin==0.55.0
pysbd==0.3.4
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.1
pytorch-metric-learning==2.9.0
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
retrying==1.4.2
rich==14.3.3
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.11.4
semver==3.0.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.3
SQLAlchemy==2.0.48
srsly==2.5.2
srt==3.5.3
submitit==1.5.4
SudachiDict-core==20260116
SudachiPy==0.6.10
sympy==1.13.1
tabulate==0.10.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.15.2
tomli==2.4.0
torch==2.5.1+cu121
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.5.1+cu121
torchcodec==0.10.0
torchmetrics==1.8.2
torchvision==0.20.1+cu121
tqdm==4.67.3
trainer==0.0.36
transformers==4.36.2
treetable==0.2.6
TTS==0.22.0
typeguard==4.5.1
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
Werkzeug==3.1.6
wrapt==2.1.1
yarl==1.23.0
zipp==3.23.0
